--- src/deepymod_torch/library_function.py
@@ -1,194 +1,0 @@
-import numpy as np
-import torch
-from torch.autograd import grad
-from itertools import combinations, product
-import torch.nn as nn
-from deepymod_torch.nn import create_deriv_data
-
-
-def library_poly(prediction, library_config):
-    '''
-    Calculates polynomials of function u up to order M of given input, including M=0. Each column corresponds to power, i.e.
-    the columns correspond to [1, u, u^2... , u^M].
-    Parameters
-    ----------
-    prediction : tensor of size (N x 1)
-        dataset whose polynomials are to be calculated.
-    library_config : dict
-        dictionary containing options for the library function.
-    Returns
-    -------
-    u : tensor of (N X (M+1))
-        Tensor containing polynomials.
-    '''
-    max_order = library_config['poly_order']
-
-    # Calculate the polynomes of u
-    u = torch.ones_like(prediction)
-    for order in np.arange(1, max_order+1):
-        u = torch.cat((u, u[:, order-1:order] * prediction), dim=1)
-
-    return u
-
-
-def library_deriv(data, prediction, library_config):
-    '''
-    Calculates derivative of function u up to order M of given input, including M=0. Each column corresponds to power, i.e.
-    the columns correspond to [1, u_x, u_xx... , u_{x^M}].
-    Parameters
-    ----------
-    data : tensor of size (N x 2)
-        coordinates to whose respect the derivatives of prediction are calculated. First column is time, space second column.
-    prediction : tensor of size (N x 1)
-        dataset whose derivatives are to be calculated.
-    library_config : dict
-        dictionary containing options for the library function.
-    Returns
-    -------
-    time_deriv: tensor of size (N x 1)
-        First temporal derivative of prediction.
-    u : tensor of (N X (M+1))
-        Tensor containing derivatives.
-    '''
-    max_order = library_config['diff_order']
-    
-    dy = grad(prediction, data, grad_outputs=torch.ones_like(prediction), create_graph=True)[0]
-    time_deriv = dy[:, 0:1]
-    
-    if max_order == 0:
-        du = torch.ones_like(time_deriv)
-    else:
-        du = torch.cat((torch.ones_like(time_deriv), dy[:, 1:2]), dim=1)
-        if max_order >1:
-            for order in np.arange(1, max_order):
-                du = torch.cat((du, grad(du[:, order:order+1], data, grad_outputs=torch.ones_like(prediction), create_graph=True)[0][:, 1:2]), dim=1)
-
-    return time_deriv, du
-
-
-def library_1D_in(data, prediction, library_config):
-    '''
-    Calculates a library function for a 1D+1 input for M coupled equations consisting of all polynomials up to order K and derivatives up to order
-    L and all possible combinations (i.e. combining two terms) of these.
-
-    Parameters
-    ----------
-    data : tensor of size (N x 2)
-        coordinates to whose respect the derivatives of prediction are calculated. First column is time, space second column.
-    prediction : tensor of size (N x M)
-        dataset from which the library is constructed.
-    library_config : dict
-        dictionary containing options for the library function.
-
-    Returns
-    -------
-    time_deriv_list : tensor list of length of M
-        list containing the time derivatives, each entry corresponds to an equation.
-    theta : tensor
-        library matrix tensor.
-    '''
-    poly_list = []
-    deriv_list = []
-    time_deriv_list = []
-
-    # Creating lists for all outputs
-    for output in torch.arange(prediction.shape[1]):
-        time_deriv, du = library_deriv(data, prediction[:, output:output+1], library_config)
-        u = library_poly(prediction[:, output:output+1], library_config)
-
-        poly_list.append(u)
-        deriv_list.append(du)
-        time_deriv_list.append(time_deriv)
-
-    samples = time_deriv_list[0].shape[0]
-    total_terms = poly_list[0].shape[1] * deriv_list[0].shape[1]
-
-    # Calculating theta
-    if len(poly_list) == 1:
-        theta = torch.matmul(poly_list[0][:, :, None], deriv_list[0][:, None, :]).view(samples, total_terms) # If we have a single output, we simply calculate and flatten matrix product between polynomials and derivatives to get library
-    else:
-        theta_uv = torch.cat([torch.matmul(u[:, :, None], v[:, None, :]).view(samples, total_terms) for u, v in combinations(poly_list, 2)], 1)  # calculate all unique combinations between polynomials
-        theta_dudv = torch.cat([torch.matmul(du[:, :, None], dv[:, None, :]).view(samples, total_terms)[:, 1:] for du, dv in combinations(deriv_list, 2)], 1) # calculate all unique combinations of derivatives
-        theta_udu = torch.cat([torch.matmul(u[:, 1:, None], du[:, None, 1:]).view(samples, (poly_list[0].shape[1]-1) * (deriv_list[0].shape[1]-1)) for u, dv in product(poly_list, deriv_list)], 1)  # calculate all unique products of polynomials and derivatives
-        theta = torch.cat([theta_uv, theta_dudv, theta_udu], dim=1)
-
-    return time_deriv_list, theta
-
-
-def library_new(prediction, library_config):
-    X, dX = prediction
-    dt = dX[:, 0, 0:1, 0]
-    dx = dX[:, :, 1, 0]
-
-    u = library_poly(X, library_config)[:, :, None]
-    du = torch.cat((torch.ones((dx.shape[0], 1)), dx), dim=1)[:, None, :]
-    theta = (u @ du).reshape(u.shape[0], -1)
-    return [dt], theta
-
-
-class Library(nn.Module):
-    def __init__(self, input_dim, output_dim, diff_order, poly_order):
-        super().__init__()
-        self.poly_order = poly_order
-        self.diff_order = diff_order
-        self.sparsity_mask_list = self.sparsity_mask_constructor(input_dim, output_dim, self.diff_order)
-
-    def forward(self, input):
-        dt, theta = self.theta(input)
-        sparse_theta_list = [theta[:, sparsity_mask] for sparsity_mask in self.sparsity_mask_list]
-        return input[0], dt, sparse_theta_list
-
-    def sparsity_mask_constructor(self, input_dim, output_dim, max_order):
-        sample_data = (torch.ones((1, output_dim), dtype=torch.float32), torch.ones((1, max_order, input_dim, output_dim), dtype=torch.float32)) # we run a single forward pass on fake data to infer shapes
-        total_terms = self.theta(sample_data)[1].shape[1]
-        sparsity_mask_list = [torch.arange(total_terms) for _ in torch.arange(output_dim)]
-
-        return sparsity_mask_list
-
-    def theta(self, input):
-        X, dX = input
-        dt = dX[:, 0, 0:1, 0]
-        dx = dX[:, :, 1, 0]
-
-        # Calculate the polynomes of u
-        u = torch.ones_like(X)
-        for order in np.arange(1, self.poly_order+1):
-            u = torch.cat((u, u[:, order-1:order] * X), dim=1)
-
-        # Calculate derivs
-        du = torch.cat((torch.ones((dx.shape[0], 1)), dx), dim=1)
-
-        theta = (u[:, :, None] @ du[:, None, :]).reshape(u.shape[0], -1)
-
-        return [dt], theta
-
-def library_2Din_1Dout(data, prediction, library_config):
-        '''
-        Constructs a library graph in 1D. Library config is dictionary with required terms.
-        '''
-
-        # Polynomial
-        
-        max_order = library_config['poly_order']
-        u = torch.ones_like(prediction)
-
-        for order in np.arange(1, max_order+1):
-            u = torch.cat((u, u[:, order-1:order] * prediction), dim=1)
-
-        # Gradients
-        du = grad(prediction, data, grad_outputs=torch.ones_like(prediction), create_graph=True)[0]
-        u_t = du[:, 0:1]
-        u_x = du[:, 1:2]
-        u_y = du[:, 2:3]
-        du2 = grad(u_x, data, grad_outputs=torch.ones_like(prediction), create_graph=True)[0]
-        u_xx = du2[:, 1:2]
-        u_xy = du2[:, 2:3]
-        u_yy = grad(u_y, data, grad_outputs=torch.ones_like(prediction), create_graph=True)[0][:, 2:3]
- 
-        du = torch.cat((torch.ones_like(u_x), u_x, u_y , u_xx, u_yy, u_xy), dim=1)
-
-        samples= du.shape[0]
-        # Bringing it together
-        theta = torch.matmul(u[:, :, None], du[:, None, :]).view(samples,-1)
-        
-        return [u_t], theta
