+++ src/deepymod_torch/network.py
@@ -1,19 +1,7 @@
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from functools import reduce
 
-
-def create_deriv_data(X, max_order):
-    ''' Utility function to create the deriv object'''
-    if max_order == 1:
-        dX = (torch.eye(X.shape[1]) * torch.ones(X.shape[0])[:, None, None])[:, None, :]
-    else: 
-        dX = [torch.eye(X.shape[1]) * torch.ones(X.shape[0])[:, None, None]]
-        dX.extend([torch.zeros_like(dX[0]) for order in range(max_order-1)])
-        dX = torch.stack(dX, dim=1)
-        
-    return (X, dX)
 
 class Linear(nn.Linear):
     ''' Linear layer completely similar to Pytorch's; only takes in deriv data tuple and also outputs it.'''
@@ -55,4 +43,23 @@ class Tanh(ActivationFunction):
                                        lambda ds, x: 1 / torch.cosh(x)**2,
                                        lambda ds, x: -2 * ds[0] * ds[1],
                                        lambda ds, x: ds[2]**2 / ds[1] - 2* ds[1]**2] # ordered list of activation function and its derivatives
-       
+
+
+class Library(nn.Module):
+    def __init__(self, input_dim, output_dim, diff_order):
+        super().__init__()
+        self.diff_order = diff_order
+        self.total_terms = self.terms(input_dim, output_dim, self.diff_order)
+        self.sparsity_mask_list = [torch.arange(self.total_terms) for _ in torch.arange(output_dim)]
+        self.coeff_vector_list = nn.ParameterList([torch.nn.Parameter(torch.rand((self.total_terms, 1), dtype=torch.float32)) for _ in torch.arange(output_dim)])
+
+    def forward(self, input):
+        dt, theta = self.theta(input)
+        sparse_theta_list = [theta[:, sparsity_mask] for sparsity_mask in self.sparsity_mask_list]
+        return input, dt, sparse_theta_list, self.coeff_vector_list
+
+    def terms(self, input_dim, output_dim, max_order):
+        sample_data = (torch.ones((1, output_dim), dtype=torch.float32), torch.ones((1, max_order, input_dim, output_dim), dtype=torch.float32)) # we run a single forward pass on fake data to infer shapes
+        total_terms = self.theta(sample_data)[1].shape[1]
+
+        return total_terms
