+++ src/deepymod_torch/training.py
@@ -1,58 +1,36 @@
 import torch
-import time, sys
-from torch.utils.tensorboard import SummaryWriter
+import time
 
+from deepymod_torch.output import Tensorboard, progress
+from deepymod_torch.losses import reg_loss, mse_loss, l1_loss
 from deepymod_torch.sparsity import scaling
-from deepymod_torch.tensorboard import custom_board
-from deepymod_torch.losses import RegLoss, MSELoss, L1Loss, DeepMoDLoss
 
-
-def train(data, target, model, optimizer, max_iterations, loss_func=DeepMoDLoss, loss_func_args={'l1': 10e-5}):
+def train(data, target, model, optimizer, max_iterations, loss_func_args):
     start_time = time.time()
-    loss_func = loss_func(**loss_func_args)
-    
+    number_of_terms = [coeff_vec.shape[0] for coeff_vec in model[-1].coeff_vector_list]
+    board = Tensorboard(number_of_terms)
+
     # Training
     print('| Iteration | Progress | Time remaining |     Cost |      MSE |      Reg |       L1 |')
-    for iteration in torch.arange(1, max_iterations + 1):
+    for iteration in torch.arange(0, max_iterations + 1):
         # Calculating prediction and library and scaling
-        output = model(data)  # prediction, time_deriv_list, theta
-        loss = loss_func(output, target, model.coeff_vector_list)
-        
+        prediction, time_deriv_list, sparse_theta_list, coeff_vector_list = model(data)
+        coeff_vector_scaled_list = [scaling(coeff_vector, sparse_theta, time_deriv) for time_deriv, sparse_theta, coeff_vector in zip(time_deriv_list, sparse_theta_list, coeff_vector_list)]
+
+        # Calculating loss
+        loss_reg = reg_loss(time_deriv_list, sparse_theta_list, coeff_vector_list)
+        loss_mse = mse_loss(prediction[0], target)
+        loss_l1 = l1_loss(coeff_vector_scaled_list, loss_func_args['l1'])
+        loss = torch.sum(loss_reg) + torch.sum(loss_mse) + torch.sum(loss_l1)
+
+        # Writing
+        if iteration % 100 == 0:
+            progress(iteration, start_time, max_iterations, loss.item(), torch.sum(loss_mse).item(), torch.sum(loss_reg).item(), torch.sum(loss_l1).item())
+            board.write(iteration, loss, loss_mse, loss_reg, loss_l1, coeff_vector_list, coeff_vector_scaled_list)
+
         # Optimizer step
         optimizer.zero_grad()
         loss.backward()
         optimizer.step()
+    board.close()
 
-        if iteration % 50 == 0:
-            elapsed_time = time.time() - start_time
-            progress(iteration, max_iterations, elapsed_time, loss.item(), 0, 0, 0)
-
-
-def progress(iteration, max_iteration, elapsed_time, cost, MSE, PI, L1):
-    percent = iteration.item()/max_iteration * 100
-    time_left = elapsed_time * (max_iteration/iteration - 1) if iteration != 0 else 0
-    sys.stdout.write(f"\r  {iteration:>9}   {percent:>7.2f}%   {time_left:>13.0f}s   {cost:>8.2e}   {MSE:>8.2e}   {PI:>8.2e}   {L1:>8.2e} ")
-    sys.stdout.flush()
-
-'''
-    # preparing tensorboard writer
-    writer = SummaryWriter()
-    writer.add_custom_scalars(custom_board(model.coeff_vector_list))
-# Tensorboard stuff
-            writer.add_scalar('Total loss', loss, iteration)
-            for idx in torch.arange(len(MSE_cost_list)):
-                # Costs
-                writer.add_scalar('MSE '+str(idx), MSE_cost_list[idx], iteration)
-                writer.add_scalar('Regression '+str(idx), reg_cost_list[idx], iteration)
-                writer.add_scalar('L1 '+str(idx), l1_cost_list[idx], iteration)
-
-                # Coefficients
-                for element_idx, element in enumerate(torch.unbind(model.coeff_vector_list[idx])):
-                    writer.add_scalar('coeff ' + str(idx) + ' ' + str(element_idx), element, iteration)
-
-                # Scaled coefficients
-                for element_idx, element in enumerate(torch.unbind(coeff_vector_scaled_list[idx])):
-                    writer.add_scalar('scaled_coeff ' + str(idx) + ' ' + str(element_idx), element, iteration)
-
-    writer.close()
-'''
